Fix/styles call (#11)
* feat: :construction: callStyles-fixedRemoveEventListenerBug

* feat: :recycle: styles-n-refactor-4-firstDeploy
