+++ src/components/Call/InputJoinCall/InputJoinCall.jsx
@@ -6,7 +6,7 @@ function InputJoinCall({ joinCall, setRemoteCallID }) {
   };
   return (
     <div
-      className="absolute inset-0 z-10  flex flex-col
+      className="fixed inset-0 z-50  flex flex-col
     items-center justify-center gap-5 p-10 drop-shadow-lg backdrop-blur-sm"
     >
       <div className="flex ">
