+++ src/router/index.jsx
@@ -1,8 +1,8 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import { createBrowserRouter } from 'react-router-dom';
 import Root from '../layout/Root';
-import App from '../App';
-import PeerToPeer from '../components/PeerToPeer';
+import Home from '../pages/Home';
+import PeerToPeer from '../components/Call/Call';
 import NotFound from '../pages/NotFound';
 
 const router = createBrowserRouter([
@@ -14,16 +14,12 @@ const router = createBrowserRouter([
     children: [
       {
         index: true,
-        element: <App />,
+        element: <Home />,
       },
       {
         path: '/call',
         element: <PeerToPeer />,
       },
-      // {
-      //   path: '/about',
-      //   element: <App />,
-      // },
     ],
   },
 ]);
